requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-redis==6.0.0
djangorestframework==3.16.1
drf-yasg==1.21.10
easyocr==1.7.2
filelock==3.19.1
fsspec==2025.7.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
inflection==0.5.1
Jinja2==3.1.6
kaitaistruct==0.10
kombu==5.5.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.22
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.1
selenium==4.35.0
selenium-wire==5.1.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.6.11
torch==2.8.0
torchvision==0.23.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zstandard==0.24.0
